Remove references to linear precision from GIMP UI
As a babl patch has been applied to disable the babl flips, GIMP linear
and gamma precisions produce the same results. This patch removes UI
references to linear precision to simplify the UI, preserves
compatibility of XCF files with default GIMP, and allows to open linear
precision XCF files from the PhotoFlow plug-in.