Remove wrap_in_gamma_cast from plug_in_compat(pdb)
This function flips pixels when the babl flips are enabled,
possibly in response to gamma hack settings. Either way, with
the babl flips disabled this function doesn't do anything.

Not sure if plug-in-compat-cmds.c also needs to be modified,
but even though this file auto-generated the old code
keeps reappearing if this file isn't also modified.