FG/BG Color dialog is color-managed
But the color picker and the Toolbox Show foreground &
background still show the wrong colors, except for images
in the same color space as the original image from which
colors were first picked. For images in other color spaces
the numbers are correct but the displayed color is wrong.

The color pickers only produce correct LCH values for images
in well-behaved linear gamma ICC profile color spaces.