Remove printf statement from gimpimage-color-profile.c
and comment out line that checks to see if the image has an
assigned profile in functon gimp_image_get_color_profile. The default
GIMP code doesn't make this check, but I'm not sure if the patched
GIMP function needs the line or not, as the profile is used to
retrieve the colorants.