Changes to make the 20170323 merge with upstream compile and run
properly. At this point the legacy blend modes are still in the
code, but I don't think they are being used anywhere. And the
layer blend UI has a non-functional element that in default GIMP
allows the user to choose between "Default" and "Legacy". This
particular default GIMP user choice is based on default GIMP
code that transforms the RGB channel encoding between linear RGB
and RGB encoded using the sRGB TRC, which code is not part of
CCE. In CCE the image's RGB color space profile's TRC determines
the RGB channel encoding.